demo: rename system-service-monitor -> system-services-monitor and drop DCGM duplicates
Per @deesharma24 review on #891 (Apr 13 / Apr 28):

1. Rename demo and health-monitor to plural system-services-monitor to
   reflect that the module can monitor more than one service (FM +
   nvidia-persistenced today, more systemd units in the future). Renames:
     demos/system-service-monitor -> demos/system-services-monitor
     health-monitors/system-service-monitor -> health-monitors/system-services-monitor
     system_service_monitor (Python package) -> system_services_monitor
   All K8s manifests, pyproject.toml, Dockerfile, logger, CLI entry
   points and imports updated. ADR file retains original filename as a
   historical record of the scope decision.

2. Drop DCGM-overlapping checks from demos/system-services-monitor/ to
   match the same cleanup already done in the core health-monitors
   package in commit d8ccacba:
     - checks/pcie_check.py        (DCGM_HEALTH_WATCH_PCIE)
     - checks/clock_check.py       (DCGM_FI_DEV_CLOCK_THROTTLE_REASONS)
     - checks/fabric_check.py      (DCGM exporter NVLink HTTP scraping)
     - tests/test_pcie_check.py
     - nv-hostengine from DEFAULT_GPU_SERVICES (covered by
       GpuDcgmConnectivityFailure in gpu-health-monitor)
   The demo now keeps only non-DCGM-visible signals: ServiceChecker
   (systemd), CUDAValidator, plus the FabricStateChecker design documented
   in the README. requests dependency dropped (was only used by the
   deleted NVLink check).

3. Rebased onto current upstream/main to resolve the BEHIND state.

All commits DCO-signed.
