+++ demos/README.md
@@ -73,7 +73,7 @@ Run these locally on your laptop — no GPU hardware needed.
 
 **Best for:** Understanding how NVSentinel's node-drainer can delegate pod eviction to external controllers for custom drain workflows coordinated with HPC schedulers.
 
-### [System Service Monitor](system-service-monitor/)
+### [System Service Monitor](system-services-monitor/)
 
 **What it shows:** Standalone DaemonSet that detects Fabric Manager failures, per-GPU fabric state issues, and CUDA context failures -- infrastructure problems invisible to DCGM-based monitoring.
 
@@ -83,7 +83,7 @@ Run these locally on your laptop — no GPU hardware needed.
 
 **Related issue:** [#883](https://github.com/NVIDIA/NVSentinel/issues/883)
 
-**Note:** For native NVSentinel integration (gRPC HealthEvents to platform-connector), see [`health-monitors/system-service-monitor/`](../health-monitors/system-service-monitor/).
+**Note:** For native NVSentinel integration (gRPC HealthEvents to platform-connector), see [`health-monitors/system-services-monitor/`](../health-monitors/system-services-monitor/).
 
 
 ### [Local Custom Remediation Demo](local-custom-remediation-demo/)
