+++ demos/system-services-monitor/Dockerfile
