+++ demos/system-services-monitor/k8s/rbac.yaml
@@ -1,10 +1,10 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: system-service-monitor
+  name: system-services-monitor
   namespace: nvsentinel
   labels:
-    app.kubernetes.io/name: system-service-monitor
+    app.kubernetes.io/name: system-services-monitor
     app.kubernetes.io/instance: nvsentinel
     app.kubernetes.io/component: metrics
 
@@ -12,9 +12,9 @@ metadata:
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
-  name: system-service-monitor
+  name: system-services-monitor
   labels:
-    app.kubernetes.io/name: system-service-monitor
+    app.kubernetes.io/name: system-services-monitor
     app.kubernetes.io/instance: nvsentinel
     app.kubernetes.io/component: metrics
 rules:
@@ -26,16 +26,16 @@ rules:
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
 metadata:
-  name: system-service-monitor
+  name: system-services-monitor
   labels:
-    app.kubernetes.io/name: system-service-monitor
+    app.kubernetes.io/name: system-services-monitor
     app.kubernetes.io/instance: nvsentinel
     app.kubernetes.io/component: metrics
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: system-service-monitor
+  name: system-services-monitor
 subjects:
   - kind: ServiceAccount
-    name: system-service-monitor
+    name: system-services-monitor
     namespace: nvsentinel
