+++ demos/system-services-monitor/requirements.txt
@@ -1,3 +1,2 @@
 prometheus_client>=0.20.0
-requests>=2.32.2
 pyyaml>=6.0
