+++ demos/system-services-monitor/tests/test_monitor.py
@@ -1,14 +1,14 @@
 """Tests for the main monitor module and config."""
 
 import os
-from unittest.mock import patch, MagicMock
+from unittest.mock import patch
 
 import sys
 
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), ".."))
 
 from config import MonitorConfig
-from monitor import FabricManagerMonitor
+from monitor import SystemServicesMonitor
 
 
 class TestMonitorConfig:
@@ -21,8 +21,10 @@ class TestMonitorConfig:
         assert config.boot_grace_period == 300
         assert config.enable_fabric_check is True
         assert config.enable_cuda_validation is False
-        assert config.clock_throttle_ratio == 0.85
         assert "nvidia-fabricmanager" in config.gpu_services
+        # nv-hostengine is covered by gpu-health-monitor / DCGM and must
+        # not appear in the default demo service list.
+        assert "nv-hostengine" not in config.gpu_services
 
     def test_from_env(self):
         env = {
@@ -34,13 +36,8 @@ class TestMonitorConfig:
             "FLAP_WINDOW": "300",
             "FLAP_THRESHOLD": "5",
             "ENABLE_FABRIC_CHECK": "true",
-            "ENABLE_PCIE_CHECK": "false",
-            "ENABLE_CLOCK_CHECK": "true",
-            "ENABLE_NVLINK_CHECK": "false",
             "ENABLE_CUDA_VALIDATION": "true",
             "CUDA_VALIDATION_INTERVAL": "1200",
-            "DCGM_EXPORTER_URL": "http://dcgm:9400",
-            "CLOCK_THROTTLE_RATIO": "0.90",
         }
         with patch.dict(os.environ, env, clear=False):
             config = MonitorConfig.from_env()
@@ -51,12 +48,8 @@ class TestMonitorConfig:
         assert config.node_name == "test-node"
         assert config.boot_grace_period == 120
         assert config.flap_threshold == 5
-        assert config.enable_pcie_check is False
-        assert config.enable_nvlink_check is False
         assert config.enable_cuda_validation is True
         assert config.cuda_validation_interval == 1200
-        assert config.dcgm_exporter_url == "http://dcgm:9400"
-        assert config.clock_throttle_ratio == 0.90
 
     def test_custom_gpu_services(self):
         with patch.dict(os.environ, {"GPU_SERVICES": "svc-a, svc-b , svc-c"}, clear=False):
@@ -75,7 +68,7 @@ class TestMonitorConfig:
                 assert config.enable_cuda_validation is False
 
 
-class TestFabricManagerMonitor:
+class TestSystemServicesMonitor:
     """Tests for the monitor orchestrator."""
 
     def _make_monitor(self, **overrides):
@@ -85,20 +78,14 @@ class TestFabricManagerMonitor:
             "node_name": "test-node",
             "boot_grace_period": 0,  # no grace period in tests
             "enable_fabric_check": True,
-            "enable_pcie_check": False,
-            "enable_clock_check": False,
-            "enable_nvlink_check": False,
             "enable_cuda_validation": False,
         }
         defaults.update(overrides)
         config = MonitorConfig(**defaults)
-        return FabricManagerMonitor(config)
+        return SystemServicesMonitor(config)
 
-    @patch("monitor.NVLinkFabricChecker")
-    @patch("monitor.ClockChecker")
-    @patch("monitor.PCIeChecker")
     @patch("monitor.ServiceChecker")
-    def test_check_cycle_all_healthy(self, MockSvc, MockPCIe, MockClk, MockNVL):
+    def test_check_cycle_all_healthy(self, MockSvc):
         from checks.service_check import FabricManagerStatus, ServiceStatus
 
         mock_svc = MockSvc.return_value
@@ -108,22 +95,18 @@ class TestFabricManagerMonitor:
         mock_svc.check_all_gpu_services.return_value = {
             "nvidia-fabricmanager": ServiceStatus(name="nvidia-fabricmanager", active=True),
             "nvidia-persistenced": ServiceStatus(name="nvidia-persistenced", active=True),
-            "nv-hostengine": ServiceStatus(name="nv-hostengine", active=True),
         }
 
         monitor = self._make_monitor()
         monitor._service_checker = mock_svc
         monitor.run_check_cycle()
 
-        # Verify fabric_manager_up was called (metrics are global singletons)
+        # Verify service checks were invoked
         mock_svc.check_fabric_manager.assert_called_once()
         mock_svc.check_all_gpu_services.assert_called_once()
 
-    @patch("monitor.NVLinkFabricChecker")
-    @patch("monitor.ClockChecker")
-    @patch("monitor.PCIeChecker")
     @patch("monitor.ServiceChecker")
-    def test_check_cycle_fabric_manager_down(self, MockSvc, MockPCIe, MockClk, MockNVL):
+    def test_check_cycle_fabric_manager_down(self, MockSvc):
         from checks.service_check import FabricManagerStatus, ServiceStatus
 
         mock_svc = MockSvc.return_value
