+++ health-monitors/system-services-monitor/system_services_monitor/protos/__init__.py
